Update tfswitch
v0.8.832 of tfswitch can now also read from .terraform-version files, so we
don't need to implement that ourselves.
It has also learned how to read the version from the required_version
constraint in the terraform config itself.

tfswitch likes to think it runs interactively, so the output get cleaned up a little bit.