Set a lock-timeout
This will keep retrying to obtain the state lock for 5 minutes before failing.
Current behaviour is to fail immediately, and the only sensible action is to retry it manually.

5 minutes should be long enough that it's quicker than noticing and retrying manually, but not long enough
that builds are stuck if the state lock is stale.

It's not exposed as an option, but it should address the issue seen #113