Fix empty plan in PR comments with terraform 0.14
With Terraform 0.14, the plan output no longer has the `----`
separator line between the boring bit and the interesting bit
of the plan (There is no boring bit).

This means we could get rid of the `sed` line completely if
we only used terraform 0.14. If we want to work with versions
both pre and post 0.14, we need to be a bit smarter.