Add Terraform 1.0.X compatible check
'An execution plan....' does not appear in the output of Terraform plan/apply
for version 1.0.1 — this means that the simple examples of a plan-on-branch,
apply-on-main flow have high chances of never working:

The full plan output will be used for an exact string comparison, and it just
so happens that resource states can be refreshed in different orders. This means
that if comparing the two, they will often not match, despite the resulting
plan being completely identical in effect.