Add Helm CI orb (#370)
This adds a CI orb for Helm.

The orb itself is quite focused on the standard KMI use-case - e.g. it assumes you will have
values files associated with the chart and no values being passed in directly in the helm
templating step. I think this is acceptable for a first version of the orb and we can add
parameterisation for other use cases in future.