Update to using the Jukie repo for tfswitch in terraform docker container
There is a current issue with the original tfswitch as hashicorp changed
their website and so the bit of tfswitch which scrapes that info is
failing. This is a temporary fix until the PR is merged on tfswitch.

github issue: https://github.com/warrensbox/terraform-switcher/pull/225
differences between versions: https://github.com/warrensbox/terraform-switcher/compare/master...Jukie:fix-release-downloads