Adds kmi-region param to deploy job
Currently, the EU1 UAT environment works as a development environment, very different to the OEA-owned AP1 UAT (Sandbox). Because of this, on a merge to main we will want to release automatically to EU1 UAT but not AP1 UAT, hence the introduction of a kmi-region parameter that allows us to release on specific regions only.